+++ README.md
@@ -5,7 +5,7 @@ This script tries to make more accesible and help automate several certificate r
 requirements:
 `pyOpenSSL`
 
-## crl_check
+## check_crl
 crl_check is a nagios friendly script to check the expiration date of crls and alert when 30, 15, and 3 days are left.
 you can try it on a terminal running 
-`./crl_check.py path/to/crl.crl`
+`./check_crl.py path/to/crl.crl`
