+++ testcert.cer
@@ -1,0 +1,34 @@
+-----BEGIN CERTIFICATE-----
+MIIHFzCCBP+gAwIBAgITALZSI6esQ6CC19MXUxJx7h1oXTANBgkqhkiG9w0BAQsFADBaMTowOAYD
+VQQDDDFBdXRvcmlkYWQgQ2VydGlmaWNhZG9yYSBSYcOteiBOYWNpb25hbCBkZSBVcnVndWF5MQ8w
+DQYDVQQKEwZBR0VTSUMxCzAJBgNVBAYTAlVZMB4XDTE0MTAwMTE3MzM1MFoXDTMxMTAyNzE3MzM1
+MFowbTE8MDoGA1UEAxMzQXV0b3JpZGFkIENlcnRpZmljYWRvcmEgZGVsIE1pbmlzdGVyaW8gZGVs
+IEludGVyaW9yMSAwHgYDVQQKExdNaW5pc3RlcmlvIGRlbCBJbnRlcmlvcjELMAkGA1UEBhMCVVkw
+ggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQDOA7Q5wAmR8atkuhb6Gxp4GWizWSD3askh
+KR6QxXJPRaBU7/NLRVBOi1dmlRogo9mumpvGFQT2Td7Tfvv4rHiWh2QY/yI0n9XXzmotpYItyrBZ
+L0xeAJ4sTd9Q63hd0Hgbl8EwM5HxiWAa2FZQMvDK00qG/NyPEJ7gp2kBC+rcbZyaM17IbZWjJVhQ
++vZufEYNJcqjvmfj+aRZhkUOQYagfhxakE6LtT/OEo8TbTLNRGFM2in3lhULKPdJKqXCyX+fhwYm
+z5ANNzG1Fct3poj6sjiqfWmoJxr9TBwvdJAYt1TgzQcKZKhTMxVp2zOGaQBm5DmbgqVhFMCsPDmE
+5qCgqAzFkSLTJ2dBELXPdejySdOa1jLoeLP7y+1SMo4/ko23cuRX9fMiQKljJab6PGXKKfl4OOgA
+ySHuH6lC1dChQTusRi5J7VuKde1sRyrtlLjJ6Jn1TFLMZHAjuUr9GVXGGwD+4CkCemO6lhRLoV7o
+eXbMjHhIpyAIImxMdXVg66UoSGNCGNgkAukpyAtHvXE+ccRjqrFB2+x632syuznYzQs53mVbRf0j
+zR7Dg8ea7z2Q+2AaBalN8U9bgYZI4aSPzejGOwbnwr2XEuA3iiK3RfXMESAcO4mrkNB0KAOqocAV
+CWNtQA1opEXDd7xEvKmRRBA4rNUZdHT00xtTKj7D5QIDAQABo4IBwTCCAb0wQgYIKwYBBQUHAQEE
+NjA0MDIGCCsGAQUFBzAChiZodHRwOi8vd3d3LmFnZXNpYy5ndWIudXkvYWNybi9hY3JuLmNlcjAO
+BgNVHQ8BAf8EBAMCAQYwEgYDVR0TAQH/BAgwBgEB/wIBADBiBgNVHR8EWzBZMCygKqAohiZodHRw
+Oi8vd3d3LmFnZXNpYy5ndWIudXkvYWNybi9hY3JuLmNybDApoCegJYYjaHR0cDovL3d3dy51Y2Uu
+Z3ViLnV5L2Fjcm4vYWNybi5jcmwwga4GA1UdIASBpjCBozBaBglghOKuHYSIBQAwTTBLBggrBgEF
+BQcCARY/aHR0cDovL3d3dy51Y2UuZ3ViLnV5L2luZm9ybWFjaW9uLXRlY25pY2EvcG9saXRpY2Fz
+L2NwX2Fjcm4ucGRmMEUGCWCE4q4dhIgFATA4MDYGCCsGAQUFBwIBFipodHRwOi8vd3d3LmFnZXNp
+Yy5ndWIudXkvYWNybi9jcHNfYWNybi5wZGYwHQYDVR0OBBYEFJ+zNDybOqlgEMrIWQRt7a1iLEj/
+MB8GA1UdIwQYMBaAFJKekbhVKD13QiwzpZhf0MmsjbWjMA0GCSqGSIb3DQEBCwUAA4ICAQBfJLta
+xuygbc5lkn/l1iPAkoaXdtCR3mx5Gvg865uAEbVgF+kAVlXV+Gk36N92JI9Evs0zirxefu3PrAbZ
+15pQNbyFdyqI/UGUj1FaXnLEX4IuBSRuGzgeu9TC7lM7e6fjIis7e/HO35wAQ8+kjerQDJmHoBP+
+skd4qFcfXPw/r2ndjXb52v2DQgH5kXaLWvFot19e9ouqUHDbe48lkmeU+xIf3QVDMV9YICSJvrY8
+oCw0JRDvA65DIkCsI4LDUv14e68E5hUm6coSdVSg5njF7x1QZUQ7MKSBRQqInWa+Lj0r8OuWWcG2
+8yR1QG9xC5/JcVBkM8YhORNCi8MLxBW7RUfuGM10zhVnTps2/brEKKkq46CbrO8cdMxWz8O6ffOR
+N2nPHkjZCmJuWCIFUvjonVHlFlsyA1lACZ9wLbIqM+AFlaguSVufpK19D7o8C1UBdMSqP8Kw31fb
+wkEtKwwXMiZYpZemI+a8c4h7M6dUC9vnufx7bRnr4ilCvySze73VfsUf+td7K/ouN5OGLWsm1Zlu
+fhkdYvmg6Mw5ITahslT4AatDNfkEb4Y7EIvDd1fBvTstWyLPwn55kufCcZr1iIKJjCJN9+erwHbC
+v4WImtOpgpWA9l2LLcjbnBKuoxCEIE9RDULvKS9ihxMZsonV1Jm0l0IkqptNLkuqN3BlVg==
+-----END CERTIFICATE-----
